requirements.txt
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
pyyaml==6.0.1
jupyter==1.0.0
ipykernel==6.27.1
tqdm==4.66.1
torch==2.1.0
torchvision==0.16.0